finished motion of pool balls for phase one

+++ Puddle/game.h
@@ -13,7 +13,7 @@ struct Game {
 	Game();
 	~Game();
 
-	void update(Uint64 deltaT);
+	void update(float deltaT);
 
 	bool generateView(ViewState viewState, ResourceManager* resManager, PuddleRenderer* renderer);
